aaalmost done...

+++ app/src/main/java/com/sanislo/nennospizza/api/data/PizzasResponse.kt
@@ -2,11 +2,10 @@ package com.sanislo.nennospizza.api.data
 
 
 import com.google.gson.annotations.SerializedName
-import com.sanislo.nennospizza.api.data.Pizza
 
 data class PizzasResponse(
-    @SerializedName("basePrice")
-    val basePrice: Int,
-    @SerializedName("pizzas")
+        @SerializedName("basePrice")
+        val basePrice: Double,
+        @SerializedName("pizzas")
     val pizzas: List<Pizza>
 )
